Fix broken test :)
